requirements.txt
# Discord Voice Bot - CPU-Optimized for Surface Pro 5
# Install with: pip install -r requirements.txt

# Discord Bot Framework
discord.py[voice]>=2.3.0
PyNaCl>=1.5.0

# LLM Integration
requests>=2.31.0
aiohttp>=3.9.0

# Speech-to-Text (Primary: Kyutai, Fallback: Whisper)
# Kyutai STT will be installed separately from https://kyutai.org/stt
# Using faster-whisper as fallback (CTranslate2 backend, 4x faster than original Whisper)
faster-whisper>=0.10.0

# Text-to-Speech (Primary: Kyutai, Fallback: Piper)
# Kyutai TTS will be installed separately
# Piper TTS as fallback - ultra-fast, CPU-optimized
piper-tts>=1.2.0

# Audio Processing
numpy>=1.24.0
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.0
sounddevice>=0.4.5  # For microphone recording (simpler and more reliable)

# Configuration
python-dotenv>=1.0.0

# Logging and Utilities
colorama>=0.4.6

# Optional: for better audio quality
# pydub>=0.25.1  # Uncomment if audio format conversion needed
